move openssl asn1 decode functions to a new module
